Deletion, notification, and bug fixes
Implemented deletion for users, groups, and goals. I created a
confirmation dialog and notification system to go with this. I changed
the database schema to automatically cascade deletion.

I collected the many little javascript files into internal and external
bundles to reduce the number of requests.

I added a footer, font, and logo.

I've added building and minifying the tailwindcss file to the docker
build and ignored the build css file.

Signed-off-by: Rick Henry <rickhenry@rickhenry.dev>