Dragging placeholder, group editing, deadline highlighting
I've added a placeholder element for items that have been dragged
out. This is just an orange box added and removed by javascript based on
the drag events. It helps keep the heights of the elements consistent
and keep them from moving around as the item is dragged. It also
re-renders the element on the server after dropping and swaps it out so
items in the completed stage don't highlight as overdue.

Overdue items now highlight with a red border and are marked for screen
readers as well. It also shows in the individual goal display.

I've adapted the group options editing to work on the groups page. It
can now be accessed from either page and returns the user to the page
they were on correctly after the changes.

Signed-off-by: Rick Henry <rickhenry@rickhenry.dev>