Better auth failure handing, small improvements
I've made various small improvements to the application in layout and
accessibilty and appearance.

I've added error emails for invalid registrations and logins, but the
pages still display as though the login attempt is succeeding. I think
this is a good balance between security and ease of use.

I also added a little Silly Ghoul easter egg

Signed-off-by: Rick Henry <rickhenry@rickhenry.dev>